Relax constraints on scroll snapping for snap-type both.
Our scroll snapping logic confined the snap position to be within
the visible snapport if snapping in both the x and y direction; however,
snaps are allowed to be outside the "snapport's search corridor" as long
as they are not too far outside (left to the user agent to determine
what constitutes "too far").  The relaxed rules are not applied if
snapping in one direction only, or if scroll-snap-stop is set to
'always'.

In addressing the issue, two other tests required tweaking since they
had no allowance for a 2D snap outside the visible snapport either
implicitly (nested-scrollIntoView-snaps.html) or explicitly
(scroll_snap_data_unittest.cc).

Bug: 626703
Change-Id: I3646ccc271d88e06f3e6655009e32ff2a9e08931