Propagate scrollbar-gutter to the viewport
The element (root or body) which propagates its overflow properties
to the viewport will also propagate its scrollbar-gutter value.

LayoutBox will not try to apply scrollbar gutters to the
aforementioned element, to avoid duplicating the ones created by
the viewport.

Note that clientWidth and clientHeight in the viewport-defining
element will return the viewport dimension excluding the size of any
rendered scrollbars, but *not* excluding empty scrollbar gutters:

  https://www.w3.org/TR/cssom-view-1/#dom-element-clientwidth
  https://www.w3.org/TR/cssom-view-1/#dom-element-clientheight

Bug: 710214
Change-Id: I65dca9791125d91b8be7b7c51779d5821f232cf9