Expand synthesized clip layer by 1 pixel
Though the clip mask layer and the masked layer are perfectly aligned in
blink and cc, the geometry of the quads may go through different mapping
code paths, depending on the types of the quads, causing slight
misalignment of the mask quad and masked quad and hairline align the
edges. Expand the mask layer by 1 pixel to ensure its coverage on the
masked layer in case of floating-point errors etc.

Bug: 1171601
Change-Id: I4b64e1b6147d4e003a1ad131a605ee4b9886d57f